validation: crypto: reduce duplicate testing in different op modes
Run time consuming algorithm combination tests only once, in async-sched,
async-plain or sync mode (in that preference order) instead of in all
the three modes, unless FULL_TEST is requested. We assume that the actual
crypto implementation being tested is the same in all the three modes so
test coverage should not suffer much.

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>
Reviewed-by: Tuomas Taipale <tuomas.taipale@nokia.com>