linux-dpdk: crypto: use devices in the order they appear in DPDK
Populate the array of crypto devices in the same order the devices
appear in DPDK (i.e. the device ID order), not in the reverse order
as done currently. This way, when selecting a device for a session,
the device with a lower ID (and which comes first in DPDK startup
parameters) is preferred.

Correspondingly, reverse the order of the null driver and openssl
driver in the test wrapper script (so that the openssl device
continues to come last when searching for a device, increasing
test coverage).

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>