linux-dpdk: crypto: do not allocate memory with no crypto devices
Do not allocate shared memory for crypto when there are no crypto
devices. Currently we do not support null crypto algorithms without
any devices even if we could and we keep it that way.

Do not print an error message if crypto capabilities are queried when
there are no enabled crypto devices as there is nothing wrong with such
capability query.

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>